moved the ABS to strategy.py and deleted the former ABS file
